migrate card out to separate location
